pyproject.toml
[tool.poetry]
name = "podping_hivewriter"
version = "0.1.0"
description = ""
authors = ["Brian of London <brian@podping.org>","Alecks Gates <alecks@podping.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
beem = "^0.24"
pyzmq = "^22.1.0"
cffi = "^1.14.5"
pydantic = "^1.8.2"
aiohttp = {extras = ["speedups"], version = "^3.7.4"}
single-source = "^0.2.0"
rfc3987 = "^1.3.8"

[tool.poetry.dev-dependencies]
black = "^21.5b2"
isort = "^5.8.0"
pytest = "^6.2.4"
pytest-cov = "^2.12.1"
pytest-asyncio = "^0.15.1"
bandit = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py37', 'py38', 'py39']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = [
    "tests",
]